[artf6165] : user-reported bug around ldap auth helper port
 A fix for exception related to ldap authHelperPort. The csvnAuthenticationProvider.getAuthHelperPort() method returns a String, as it was intended for use in writing the httpd.conf. To use for setting server.authHelperPort (as here), the value needs to be cast to Integer.
 
 * grails-app/controllers/../ServerController.groovy (editAuthentication)
 now casts the result of csvnAuthenticationProvider.getAuthHelperPort() to Integer before setting the corresponding Server property

git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2243 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f

+++ grails-app/controllers/com/collabnet/svnedge/controller/admin/ServerController.groovy
@@ -135,7 +135,7 @@ class ServerController {
         def server = Server.getServer()
         if (!server.authHelperPort) {
            server.authHelperPort =     
-                csvnAuthenticationProvider.getAuthHelperPort(server, server.ldapEnabled) 
+                csvnAuthenticationProvider.getAuthHelperPort(server, server.ldapEnabled) as Integer
         }
         return [server: server,
                 csvnConf: ConfigUtil.confDirPath(),
